+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [19.10.0] - 2026-01-16
+
+### Added
+
+- Added support for exposing additional metrics about application health contributors.
+  Enable via the `jeap.health.metric.contributor-metrics.enabled` property.
+
 ## [19.9.0] - 2026-01-14
 
 ### Changed